UserName Bug fixed. Reorganized model package.
